The ConcurrentDownloader class implements a basic concurrent downloader with progress bars. The idea is to have a task list where tasks can be scheduled by user of the class and also limits can be set, and then run you can just run the tasks.
The module includes:
1. DownloadFileInfo dataclass - stores information for scheduling tasks
   and handling errors (such as deleting unfinished downloads)
2. default_fail_handler_callback - a function that deletes files
   that have not been processed properly (by default)
3. ConcurrentDownloader - the sheriff of the town,
   manages concurrency and handle errors:
    - the number of tasks that can be downloaded concurrently
    - what to do if a task fails (ethernet can be shut down and so on)

The class has three list:
1. for scheduling
2. for running tasks,
3. for finished tasks.

The class relocates tasks between the lists based on their status,
so current tasks information is available from anywhere in the class.

But the actual tasks (coroutines/futures) are located in the task_workers list.
We pass this list to the asyncio.wait() function and refill it
when any running task is finished.

All the task info objects from running list has
their according coroutines in the task_workers.

For example:
If we have 4 workers and 2nd worker finishes, a new worker will be assigned
to its position (2nd), that insures that progress bars display properly.

When a task is finished, we use its name as the worker ID (pbar position).
This allows us to replace the worker with a new one
and maintain the correct positioning of the progress bars.

This was the most challenging part of the main downloader code,
I will write some documentation for it in the near future so
any wandering developer can use it without much pain <3.