Get rid of 💩 code, fix async limiter, exceptions.
Some parts of the code were awful.
I didn't know much about asyncio,
but now I realize how task cancellation works,
and everything is handled properly.

1. AsyncLimiter with float('inf') worked not as expected,
so I set its default "unlimited" version to the huge
number which is 1'000'000 connections per second.
It's not very neat, but anyway.

2. I completely removed the system for handling unclosed files
because it was incorrect from the beginning. The files should not
have been closed explicitly in cleanup methods,
but rather properly handled in the code itself to ensure
they are closed in case of asyncio.CancelledError.

3. Now, if we have 429 (TooManyRequests) response the task clean
things up and raise exception further, so all other tasks can
be cancelled properly.

Now everything is working pretty well. The only remaining task is
to refactor the code, as it is not currently well-structured.