New type system for aiowallhaven api
I have decided to make the API typed,
as it is much more convenient to work with it in such manner.
There is no need to remember all the JSON fields in the server response.
Moreover, these classes are easier to test, and the code itself will look cleaner.

Also, the type modules have been separated into two different files:
1. wallhaven_types.py for data classes exclusively
2. wallhaven_enums.py for enums exclusively