Add "status changed" callback for async downloader
So now users can pass some custom func to run_downloader
as a notifier (which will called automatically)

Also, it is a public method of the downloader class,
because sometimes it's handy to get status manually.

+ Unfortunately, the deque may not be more efficient than the list
because it removes elements with O(n) complexity which is
same for the list.